pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "flpoison"
version = "0.0.0"
description = "FLPoison research codebase"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }

dependencies = [
  "hdbscan>=0.8.40,<1",
  "matplotlib>=3.10,<4",
  "numpy>=2.2,<3",
  "pillow>=11.1,<12",
  "pyyaml>=6.0,<7",
  "rarfile>=4.2,<5",
  "scikit-learn>=1.6,<2",
  "scipy>=1.15,<2",
  "torch>=2.6,<3",
  "torchvision>=0.21,<1",
  "tqdm>=4.66,<5",
  "unrar>=0.4,<1",
]

[project.scripts]
flpoison = "flpoison.__main__:main"
flpoison-batch = "flpoison.cli.batchrun:main_entry"

[tool.setuptools.packages.find]
where = ["."]
include = ["flpoison*"]
